refactor(FON-468): move session attachment list and xlsx hint into transparence
