+++ apps/client/src/features/transparence/components/TableauDeBordActions.tsx
@@ -2,12 +2,12 @@ import Accordion from '@codegouvfr/react-dsfr/Accordion';
 import Badge from '@codegouvfr/react-dsfr/Badge';
 import Button from '@codegouvfr/react-dsfr/Button';
 
-import { NominationSessionAttachmentList } from '@/components/shared/NominationSessionAttachmentList';
 import { useArchivedSession } from '@/hooks/archive/useArchivedSession';
 import { useFindSessionDocsQuery } from '@queries/agenda.queries';
 import { useListNominationSessionAttachmentsQuery } from '@queries/nomination-sessions.queries';
 
 import * as importAttachments from './ImportAttachmentModal';
+import { NominationSessionAttachmentList } from './NominationSessionAttachmentList';
 import { NominationSessionDocsList } from './NominationSessionDocsList';
 import { TableauDeBordActionList } from './TableauDeBordActionsList';
